moved resume data and begin pdf creation
