+++ src/components/Resume/data.ts
@@ -1,0 +1,81 @@
+const jobs = [
+  {
+    company: "Vemo Education",
+    title: "Lead Front-End Developer",
+    startDate: "2019",
+    description:
+      "As Lead Front-End Developer at Vemo Education, I lead our team of developers as we create software for students and administrators for income share agreements. I help create and architect technical requirements and their solutions, help and mentor other developers, and review code. I also have worked on implementing a design system across projects and also built algorithms that calculate loans and income share agreements.",
+  },
+
+  {
+    company: "DeveloperTown",
+    startDate: "2013",
+    endDate: "2019",
+    title: "Developer",
+    description:
+      "At DeveloperTown, I worked as a front-end developer. My projects usually consisted of a React front-end connected to an API for data. I worked with clients to gather requirements, estimate work, and guide them through the software building process. I worked closely with our project managers, designers, and back-end developers to build awesome software.",
+  },
+
+  {
+    company: "DWA Healthcare Communications Group",
+    startDate: "2013",
+    endDate: "2013",
+    title: "Front end developer",
+    description: "At DWA, I built web interfaces for Eli Lilly.",
+  },
+
+  {
+    company: "KA+A",
+    startDate: "2012",
+    endDate: "2013",
+    title: "Interaction Designer",
+    description:
+      "At KA+A, I developed web applications for several clients, including ExactTarget (HubExchange) and Angie’s List (Big Deal).",
+  },
+
+  {
+    company: "Right On Interactive",
+    startDate: "2011",
+    endDate: "2012",
+    title: "Interactive Marketing Manager",
+    description: "adfdsaf",
+  },
+
+  {
+    company: "Kemper Technology Consulting",
+    startDate: "2010",
+    endDate: "2011",
+    title: "Web Designer/Developer",
+    description: "adfdsaf",
+  },
+
+  {
+    company: "University of Evansville",
+    startDate: "2008",
+    endDate: "2010",
+    title: "Manager of Digital Media",
+    description: "adfdsaf",
+  },
+
+  {
+    company: "Dell, Inc.",
+    startDate: "2007",
+    endDate: "2008",
+    title: "Senior Account Manager",
+    description:
+      "At Dell, Inc., I was on the phone with customers assiting them with their computer, server, and software needs. I had $1.7 million in sales (over 100% of quota every month).",
+  },
+
+  {
+    company: "Ball State University",
+    startDate: "2003",
+    endDate: "2007",
+    title: "Student",
+    description:
+      "Bachelor’s of Arts Telecommunications, Multimedia focus Minor in Marketing ",
+  },
+];
+
+const resumeData = { jobs };
+
+export default resumeData;
